# ahms-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ayurvedic Hospital Management System — Phase 1 API (Auth, Roles, Departments, Patients, Doctors, Appointments, Dashboard).",
        "title": "AHMS Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/bills": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "List bills",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by payment status (UNPAID/PARTIAL/PAID)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by bill number",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by patient ID",
                        "name": "patient_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/billing.BillResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Create a bill with line items",
                "parameters": [
                    {
                        "description": "Bill",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/billing.CreateBillRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/billing.BillResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/bills/number/{bill_no}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Get a bill by its number (BILL-2026-000001)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bill number",
                        "name": "bill_no",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/billing.BillResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/bills/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Get a bill by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/billing.BillResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/bills/{id}/payments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billing"
                ],
                "summary": "Record a payment against a bill",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/billing.PaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/billing.BillResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "billing.BillItemInput": {
            "type": "object",
            "required": [
                "description",
                "quantity",
                "rate"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "minLength": 2
                },
                "quantity": {
                    "type": "integer"
                },
                "rate": {
                    "type": "number",
                    "minimum": 0
                },
                "service_type": {
                    "type": "string"
                }
            }
        },
        "billing.BillItemResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "rate": {
                    "type": "number"
                },
                "service_type": {
                    "type": "string"
                }
            }
        },
        "billing.BillResponse": {
            "type": "object",
            "properties": {
                "bill_no": {
                    "type": "string"
                },
                "billed_by": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "due_amount": {
                    "type": "number"
                },
                "encounter_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/billing.BillItemResponse"
                    }
                },
                "net_amount": {
                    "type": "number"
                },
                "paid_amount": {
                    "type": "number"
                },
                "patient_id": {
                    "type": "string"
                },
                "patient_name": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "payments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/billing.PaymentResponse"
                    }
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
        "billing.CreateBillRequest": {
            "type": "object",
            "required": [
                "items",
                "patient_id"
            ],
            "properties": {
                "discount": {
                    "type": "number",
                    "minimum": 0
                },
                "encounter_id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/billing.BillItemInput"
                    }
                },
                "patient_id": {
                    "type": "string"
                }
            }
        },
        "billing.PaymentRequest": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "minimum": 0
                },
                "method": {
                    "type": "string",
                    "enum": [
                        "CASH",
                        "CARD",
                        "UPI",
                        "BANK_TRANSFER"
                    ]
                },
                "reference": {
                    "type": "string"
                }
            }
        },
        "billing.PaymentResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                }
            }
        },
...
        "utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}